rp2040: move flash related functions into separate file from C imports for correct LSP. Fixes #3852
